requirements.txt
efficientnet_pytorch==0.7.1
numpy==1.24.2
pandas==1.5.3
Pillow==10.1.0
plotly==5.13.1
scikit_learn==1.2.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
